requirements.txt
asgiref==3.7.2
dj-config-url==0.1.1
dj-database-url==2.1.0
Django==4.2.7
django-ckeditor==6.7.0
django-jazzmin==2.6.0
django-js-asset==2.2.0
djongo==1.3.6
dnspython==2.5.0
environ==1.0
gunicorn==21.2.0
humanize==4.9.0
packaging==23.2
pillow==10.2.0
postgres==4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pymongo==4.6.1
pypi==2.1
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.6.0
